Merge the native nupkg into the Core.Native nupkg.
There was no benefit in separating the native assets into its own package, the runtime.<rid>.PackageName strategy is for binaries to be copied to the root of project and therefore doesn't support the portable scenario where all available assets are required.

Manually author the repository element in the nupkg for SourceLink.