[Linux] Fix newline handling in help message.
The current help message for `wix` itself doesn't end in a newline. On
Windows that's only a minor oddity because command prompts are
typically preceded by a newline anyway; on Linux, they're typically
not, so it's much more unfriendly to leave the shell prompt at the end
of a line.

Also, generating newlines using Console.WriteLine instead of an
explicit `\r\n` means they automatically adapt to the local platform's
newline convention.