Handle unexpected BOM when reading from stdin in wixnative.exe
In certain cases, the encoding of stdin/stdout can be changed in
Windows to cause a BOM to be added on the first read of stdin.
To handle the unexpected BOM, the first line read from stdin is
now a "preamble" where the last character is expected to be a ":".
We use the last character because a multibyte BOM value may appear
before it.

Also updated SAL annotations to resolve IDE warnings.