Put validation tests behind #if due to confirmed MSI inconsistencies
When the validation tests are run repeatedly, they will expose an issue
in the Windows Installer where ICE validations will occasionally
fail to send all error/warning messages. The inconsistency obviously
wreaks havoc on the tests.

The tests are still interesting (and complex) enough to keep around
for manual testing. So they were placed behind an #if for easy manual
inclusion.

Note that these are "negative tests" that expose the Windows Installer
inconsistency when expected failures are missing. Other validation tests
are "positive tests" that look for success and thus will always
succeed. Therefore, the positive tests stay active.