Rework build command line for combining wixlibs and default output
Changes the command line handling for the "build" command to allow
creation of a .wixlib to accept compiled (but not linked)
intermediates. This will allow .wixlibs to combine separate build
commands (potentially with different "-arch" switches) into a single
.wixlib. This change also fixes the issue where by default output
was being written to the intermediate folder instead of the current
directory.

Fixes 6464
Fixes 6473