Rename build.cmd to devbuild.cmd to better represent its purpose
The build.cmd in the root of the repository is intended for use by
developers looking to build the full WiX Toolset. Official builds
use the appropriate build files in the src folder.

Renaming build.cmd to devbuild.cmd better represents its purpose.
As a bonus, renaming to devbuild.cmd also removes TAB completion
conflict with the .\build folder.