Create setup segment and WixAdditionalTools bundle
The setup segment is scheduled after building the shipping WiX files
to use and include the current build's tooling. The
WixAdditionalTools bundle currently only contains the ThmViewer
package but may include additional tools in the future.

At this time, the primary purpose is to produce a shipping MSI and
bundle from the WiX v4 build process and dogfood the rest of the
tooling.

Closes 4919