Allow signing targets to be overridden in simple Sdk-style .wixproj
Moved the overridable signing targets (like SignMsi) to a new
WixToolset.Signing.props that is included "at the very top of" a
.wixproj using Sdk-style project. That way these signing targets can be
overridden in the .wixproj. When the overridable signing targets were
defined in WixToolset.Signing.targets, then it was necessary to use the
more complex Sdk-style Import syntax to place the signing targets low
enough in the .wixproj.

Closes 7038