Remove legacy permissions support.
In the long ago, a share could have one user/permissions pair. That's
really limited, of course, so support for _n_ user/permissions pairs
was added. In the move to WiX v4, support for that single, legacy user
was removed from the extension side but not the custom action side.
Remove that support.

Fixes https://github.com/wixtoolset/issues/issues/7632.