Introduce a new phase in the build pipeline.
A useful point in the build pipeline is after all the files in the
project have been compiled but before they've been linked.
The WiX core and extensions can operate on symbols across
the project but without operating at the source-code level.

This phase is currently named "optimize," after a moderately-similar
phase in other compiler architectures. The name is, for now, a stake in
the ground and a better alternate is welcome.