Rename inter-Burn communication related pipe file
When BootstrapperApplications move out of proc there will be another set of
pipes in Burn that behave differently from the inter-Burn communication pipes
that exist now. So renaming the existing pipe related files now to better
reflect that it is related to Burn talking to Burn when elevated or embedded.