Added some more modifications to the Sandbox. It will now prompt for configuring the Remote Debugger if it can find it on the host when doing the setup. And will prompt within the Sandbox to run the Debugger at startup if found It will also show the host-local IP address to connect to the debugger.
In addition, it will display a small menu to allow for simple selection of
a particular runtests.cmd version to execute.

I still haven't found a good way to get a debugger attached into the msiexec
environment without already having a coded breakpoint in, however.

Also added this sandbox folder to .gitignore, so that artifacts from here
aren't committed into the repo


Signed-off-by: Bevan Weiss <bevan.weiss@gmail.com>