Fix default feature to respect sections.
The bug is that components would get assigned to the default feature
even if they weren't referenced from the entry section. But without the
reference, the linker wouldn't bring them into the package. This fix
resolves references to get a "clean" list of sections, adds a
reference to the default feature, then re-resolves to pick up the
default feature from the StdLib.