Prevent unnecessary refreshes that move focus.
Spillover from having more interactive controls. In WiX v3, showing a
page moved the focus as a typical person might expect. When WiX v4
added interactivity, the same code was used to update controls (e.g.,
when checking a checkbox set a property used in a `VisibleCondition` or
`EnableCondition` for another control on the same page). This change
prevents the focus-setting behavior if the same page is being shown.

Fixes https://github.com/wixtoolset/issues/issues/8144