Ensure naked files take Subdirectory into account.
Naked files generated their ids using the attributes that are common to
both naked and clothed files. But naked files also support @Subdirectory
to magic up a subdirectory in a specified directory (@Directory) or the
default INSTALLFOLDER. That subdirectory needs to factor in to the
generated file id (which is then used as the component id too).
Without it, generated ids for files with the same name but from
different @Subdirectory values would be duplicated. (Authored file ids
must also continue to be supported.)

Naked files now generate different file and component ids. :(

Fixes https://github.com/wixtoolset/issues/issues/8674